+++ docs/en/changelog.rst
@@ -21,6 +21,8 @@ Core
 
 * Can add custom data to IkParameterization that is also passed to the ik filters. :meth:`.IkParameterization.SetCustomData`
 
+* Can use IkParameterization to :meth:`.Robot.Manipulator.CheckEndEffectorCollision` even if it isn't a Transform6D type.
+
 Viewer
 ------